(PC-32634) feat(venue): setup venue image preview
